shop1/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.2
django-ckeditor==6.0.0
django-environ==0.4.5
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-registration-redux==2.9
django-widget-tweaks==1.4.8
Pillow==8.2.0
pip==21.0.1
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
setuptools==52.0.0
six==1.16.0
sqlparse==0.4.1
wheel==0.36.2
wincertstore==0.2
